requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process
pbr!=2.1.0,>=2.0.0 # Apache-2.0
github-webhook>=1.0.2 # Apache-2.0
asv>=0.3 # BSD
GitPython>=2.1.11 # BSD
